Split stock tests, add missing given, when, then in warehouse tests
