+++ saleor/graphql/warehouse/tests/mutations/test_warehouse_shipping_zone_unassign.py
@@ -17,6 +17,7 @@ mutation unassignWarehouseShippingZone($id: ID!, $shippingZoneIds: [ID!]!) {
 def test_shipping_zone_unassign_from_warehouse(
     staff_api_client, warehouse, shipping_zone, permission_manage_products
 ):
+    # given
     assert warehouse.shipping_zones.first().pk == shipping_zone.pk
     staff_api_client.user.user_permissions.add(permission_manage_products)
     variables = {
@@ -26,9 +27,12 @@ def test_shipping_zone_unassign_from_warehouse(
         ],
     }
 
+    # when
     staff_api_client.post_graphql(
         MUTATION_UNASSIGN_SHIPPING_ZONE_WAREHOUSE, variables=variables
     )
+
+    # then
     warehouse.refresh_from_db()
     shipping_zone.refresh_from_db()
     assert not warehouse.shipping_zones.all()
